requirements.txt
streamlit==1.32.0
requests==2.31.0
python-dotenv==1.0.1
moviepy==1.0.3
Pillow==10.2.0
gTTS==2.5.1
google-generativeai==0.3.2
elevenlabs==0.2.26
openai==1.12.0 